# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ScamShield AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface TextAnalysisInput {
  /** @minLength 1 */
  content: string;
  /** If true, use Explain-like-I'm-60 mode */
  simpleMode?: boolean;
}

export interface UrlAnalysisInput {
  /** @minLength 1 */
  url: string;
  simpleMode?: boolean;
}

export interface ImageAnalysisInput {
  /** Base64-encoded image data */
  imageBase64: string;
  /** MIME type of the image (e.g. image/jpeg) */
  mimeType: string;
  simpleMode?: boolean;
}

/**
 * Score from a single intelligence source
 */
export interface SignalSource {
  /** 0–100 risk score from this source */
  score?: number;
  /** Weight of this source in the ensemble (0–1) */
  weight?: number;
  /** Human-readable verdict from this source */
  verdict?: string;
  /** Whether this source was reachable and returned data */
  available?: boolean;
  /** Error message if source was unavailable */
  error?: string;
}

export type SignalBreakdownGemini = {
  score?: number;
  weight?: number;
};

export type SignalBreakdownVirustotal = SignalSource & {
  malicious?: number;
  suspicious?: number;
  harmless?: number;
  undetected?: number;
};

export type SignalBreakdownGoogleSafeBrowsing = SignalSource & {
  threats?: string[];
};

export type SignalBreakdownPhishtank = SignalSource & {
  isPhishing?: boolean;
};

export type SignalBreakdownDomainAge = SignalSource & ({
  domain?: string;
  registrationDate?: string | null;
  ageInDays?: number | null;
});

/**
 * Per-source signal breakdown for URL analyses
 */
export interface SignalBreakdown {
  gemini?: SignalBreakdownGemini;
  virustotal?: SignalBreakdownVirustotal;
  google_safe_browsing?: SignalBreakdownGoogleSafeBrowsing;
  phishtank?: SignalBreakdownPhishtank;
  domain_age?: SignalBreakdownDomainAge;
  /** Final blended ensemble score */
  ensembleScore?: number;
}

export interface Analysis {
  id: number;
  /** text | url | image */
  inputType: string;
  /** The original submitted content or URL */
  inputContent: string;
  /** 0–100 risk score (ensemble-blended for URL scans) */
  riskScore: number;
  /** e.g. Phishing, Job Scam, UPI Fraud, Investment Scam */
  scamType: string;
  /** low | medium | high | critical */
  severity: string;
  redFlags: string[];
  /** Human-readable explanation of the analysis */
  explanation: string;
  recommendations: string[];
  /** What the attacker is trying to achieve */
  attackerGoal: string;
  simpleMode?: boolean;
  /** Per-source signal breakdown (only present for URL analyses) */
  signals?: SignalBreakdown | null;
  createdAt: string;
}

export interface StatsSummary {
  totalAnalyses: number;
  highRiskCount: number;
  safeCount: number;
  averageRiskScore: number;
  scamsBlockedToday: number;
}

export interface ScamTypeCount {
  scamType: string;
  count: number;
}

export type ListAnalysesParams = {
limit?: number;
offset?: number;
};