# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ScamShield AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SignalBreakdown } from './signalBreakdown';

export interface Analysis {
  id: number;
  /** text | url | image */
  inputType: string;
  /** The original submitted content or URL */
  inputContent: string;
  /** 0–100 risk score (ensemble-blended for URL scans) */
  riskScore: number;
  /** e.g. Phishing, Job Scam, UPI Fraud, Investment Scam */
  scamType: string;
  /** low | medium | high | critical */
  severity: string;
  redFlags: string[];
  /** Human-readable explanation of the analysis */
  explanation: string;
  recommendations: string[];
  /** What the attacker is trying to achieve */
  attackerGoal: string;
  simpleMode?: boolean;
  /** Per-source signal breakdown (only present for URL analyses) */
  signals?: SignalBreakdown | null;
  createdAt: Date;
}