# lib/api-zod/src/generated/types/signalBreakdown.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ScamShield AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SignalBreakdownDomainAge } from './signalBreakdownDomainAge';
import type { SignalBreakdownGemini } from './signalBreakdownGemini';
import type { SignalBreakdownGoogleSafeBrowsing } from './signalBreakdownGoogleSafeBrowsing';
import type { SignalBreakdownPhishtank } from './signalBreakdownPhishtank';
import type { SignalBreakdownVirustotal } from './signalBreakdownVirustotal';

/**
 * Per-source signal breakdown for URL analyses
 */
export interface SignalBreakdown {
  gemini?: SignalBreakdownGemini;
  virustotal?: SignalBreakdownVirustotal;
  google_safe_browsing?: SignalBreakdownGoogleSafeBrowsing;
  phishtank?: SignalBreakdownPhishtank;
  domain_age?: SignalBreakdownDomainAge;
  /** Final blended ensemble score */
  ensembleScore?: number;
}