# lib/api-zod/src/generated/types/signalSource.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ScamShield AI API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Score from a single intelligence source
 */
export interface SignalSource {
  /** 0–100 risk score from this source */
  score?: number;
  /** Weight of this source in the ensemble (0–1) */
  weight?: number;
  /** Human-readable verdict from this source */
  verdict?: string;
  /** Whether this source was reachable and returned data */
  available?: boolean;
  /** Error message if source was unavailable */
  error?: string;
}